Embed service creation in main window
