--- DesktopApplicationTemplate.UI/Views/CreateServiceWindow.xaml
@@ -1,17 +1,0 @@
-<Window x:Class="DesktopApplicationTemplate.UI.Views.CreateServiceWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        Title="Create Service" Width="800" Height="600"
-        WindowStartupLocation="CenterOwner"
-        Style="{DynamicResource BubblyWindowStyle}">
-    <Window.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="../Themes/BubblyWindow.xaml" />
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
-    </Window.Resources>
-    <ScrollViewer VerticalScrollBarVisibility="Auto" HorizontalScrollBarVisibility="Auto">
-        <Frame x:Name="ContentFrame" NavigationUIVisibility="Hidden" />
-    </ScrollViewer>
-</Window>
